chore: separate abstract models into their own files
