+++ packages/embedder/ai_models/OpenAIGeneration.ts
@@ -4,7 +4,7 @@ import {
   GenerationModelConfig,
   GenerationModelParams,
   GenerationOptions
-} from './AbstractModel'
+} from './AbstractGenerationModel'
 
 const MAX_REQUEST_TIME_S = 3 * 60
