+++ packages/embedder/processJobQueue.ts
@@ -3,7 +3,7 @@ import ms from 'ms'
 import sleep from 'parabol-client/utils/sleep'
 import 'parabol-server/initSentry'
 import getKysely from 'parabol-server/postgres/getKysely'
-import {EmbeddingsTable} from './ai_models/AbstractModel'
+import {EmbeddingsTable} from './ai_models/AbstractEmbeddingsModel'
 import {ModelManager} from './ai_models/ModelManager'
 import {createEmbeddingTextFrom} from './indexing/createEmbeddingTextFrom'
 import {getRootDataLoader} from './indexing/getRootDataLoader'
